jake-requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
crispy-bootstrap5==2024.10; python_version >= '3.8'
dj-database-url==2.3.0
django==5.1.6; python_version >= '3.10'
django-authtools==2.0.1
django-crispy-forms==2.3; python_version >= '3.8'
django-rq==3.0.0
django-storages==1.14.4; python_version >= '3.7'
pytz==2025.1
redis==5.2.1; python_version >= '3.8'
rq==2.1.0; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
whitenoise==6.9.0; python_version >= '3.9'
